+++ src/ashley/machina_extensions/forum/views.py
@@ -14,7 +14,6 @@ from machina.core.db.models import get_model
 from machina.core.loading import get_class
 
 Forum = get_model("forum", "Forum")
-
 PermissionRequiredMixin: BasePermissionRequiredMixin = get_class(
     "forum_permission.viewmixins", "PermissionRequiredMixin"
 )