+++ src/ashley/urls.py
@@ -5,7 +5,7 @@ Ashley URLs (that includes django machina urls)
 from django.urls import include, path
 from machina import urls as machina_urls
 
-from ashley.views import ChangeUsernameView, ForumLTIView
+from ashley.views import ChangeUsernameView, ForumLTIView, ManageModeratorsView
 
 urlpatterns = [
     path("lti/forum/<uuid:uuid>", ForumLTIView.as_view(), name="forum.lti.view"),
@@ -14,5 +14,6 @@ urlpatterns = [
         ChangeUsernameView.as_view(),
         name="forum.username.change",
     ),
+    path("moderators/", ManageModeratorsView.as_view(), name="moderators"),
     path("forum/", include(machina_urls)),
 ]