+++ CHANGELOG.md
@@ -10,7 +10,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- Clean built frontend files before each build
+ - Clean built frontend files before each build
+
+### Added
+ 
+ - add new role moderator and permission to manage moderators
 
 ## [1.0.0-beta.5] - 2020-03-01